# specs/platform/api-contract.yaml
openapi: 3.1.0
...
info:
  title: FB4-App Backend
  # Dokumentversion dieser Datei nach specs/README.md Abschnitt 8 (Spec-Versionierung) —
  # unabhängig vom Wire-Versionssegment /v1 in der Server-URL unten. Ein Anstieg hier
  # begleitet jede inhaltliche Änderung dieses Vertrags; /v1 ändert sich nur bei einer
  # brechenden Schnittstellenänderung. Siehe decisions/0016-api-versionierung-und-deprecation.md.
  version: 0.3.0
  description: |
    Schreibpfade, Zwischenspeicher und Stammdaten der FB4-App.
    Aufgabenschnitt und Prinzipien: specs/platform/backend-and-api.md.
    Externe Quellsysteme: specs/platform/integrations.md.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://api.fsrfb4.de/v1
    description: Produktivbetrieb (Hetzner-VPS, FSR FB4)
...
security:
  - authentik: []
paths:
...
  /stammdaten:
    get:
      tags: [stammdaten]
      summary: Alle ferngepflegten Stammdaten in einem Aufruf
      description: |
        Erfüllt API-F-230. Die App hält das Ergebnis lokal vor (DATA, Abschnitt 4) und
        fällt bei Nichterreichbarkeit auf den im Anwendungspaket mitgelieferten
        Ausgangsbestand zurück (API-F-235).
      security: []
      responses:
        "200":
          description: Stammdatenbestand
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Stammdaten" }
        default: { $ref: "#/components/responses/Fehler" }

  # --------------------------------------------------------------------- Mensa
  /mensen:
    get:
      tags: [mensa]
      summary: Auswählbare Mensen
      description: Erfüllt API-F-230 für MENSA-F-020; Öffnungszeiten gemäß MENSA-F-047.
      security: []
      responses:
        "200":
          description: Liste der Mensen
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Mensa" }
        default: { $ref: "#/components/responses/Fehler" }
...
  /raeume:
    get:
      tags: [raum]
      summary: Vom FSR gepflegte Raumliste
      description: Erfüllt RAUM-F-100, RAUM-F-110 und RAUM-F-120.
      security: []
      responses:
        "200":
          description: Räume mit Größenklasse und E-Key-Zugänglichkeit
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Raum" }
        default: { $ref: "#/components/responses/Fehler" }
...
  /stundenplan/studiengaenge:
    get:
      tags: [stundenplan]
      summary: Rückfallliste der Studiengänge
      description: |
        Erfüllt API-F-240 und SCHED-F-254. Wird nur verwendet, wenn INT-001 nicht
        erreichbar ist; im Regelfall spricht die App INT-001 unmittelbar an (ARCH-F-020).
      security: []
      responses:
        "200":
          description: Studiengänge mit Fachsemestern
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Studiengang" }
        default: { $ref: "#/components/responses/Fehler" }
...
  /verwaltung/rollen:
    get:
      tags: [verwaltung]
      summary: Konten mit Verwaltungsrollen
      description: |
        Erfüllt ADMIN-F-070 und IDENT-F-045. Führendes System ist Authentik (INT-012);
        das Backend hält keine eigene Rollentabelle (API-F-250).
      responses:
        "200":
          description: Rollenzuweisungen
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Rollenzuweisung" }
        "401": { $ref: "#/components/responses/Unauthentifiziert" }
        "403": { $ref: "#/components/responses/Verboten" }
        default: { $ref: "#/components/responses/Fehler" }
    put:
      tags: [verwaltung]
      summary: Rollen eines Kontos setzen
      description: |
        Erfüllt ADMIN-F-070. Wirkt auf Authentik-Gruppen. Der Versuch, die letzte
        verbleibende Zuweisung der Rolle FSR-Redaktion zu entziehen, wird abgelehnt
        (ADMIN-F-080).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Rollenzuweisung" }
      responses:
        "200":
          description: Rollen gesetzt
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Rollenzuweisung" }
        "401": { $ref: "#/components/responses/Unauthentifiziert" }
        "403": { $ref: "#/components/responses/Verboten" }
        "409":
          description: Letzte Zuweisung der Rolle FSR-Redaktion (ADMIN-F-080)
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Fehler" }
        default: { $ref: "#/components/responses/Fehler" }
  /verwaltung/laufwege:
    get:
      tags: [verwaltung]
      summary: Laufwege zwischen Räumen
      description: |
        Erfüllt API-F-220 und ADMIN-F-090; Grundlage für RAUM-F-060. Der Kopf `ETag`
        trägt den Stand der Gesamtliste für die optimistische Nebenläufigkeitskontrolle
        beim Ersetzen (siehe PUT unten).
      responses:
        "200":
          description: Laufwege
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Laufweg" }
        "401": { $ref: "#/components/responses/Unauthentifiziert" }
        "403": { $ref: "#/components/responses/Verboten" }
        default: { $ref: "#/components/responses/Fehler" }
    put:
      tags: [verwaltung]
      summary: Laufwege ersetzen
      description: |
        Erfüllt ADMIN-F-090. Einträge mit unbekannter Raumkennung werden gespeichert
        und im Ergebnis als unbekannt gekennzeichnet (ADMIN-F-100). Der Kopf `If-Match`
        schützt vor dem Überschreiben einer zwischenzeitlichen Änderung durch eine
        andere Person, analog `/meldungen/entwuerfe/{entwurfId}`.
      parameters:
        - name: If-Match
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: "#/components/schemas/Laufweg" }
      responses:
        "200":
          description: Laufwege gesetzt, mit Hinweisen auf unbekannte Raumkennungen
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                required: [laufwege, unbekannteRaeume]
                properties:
                  laufwege:
                    type: array
                    items: { $ref: "#/components/schemas/Laufweg" }
                  unbekannteRaeume:
                    type: array
                    items: { type: string }
        "401": { $ref: "#/components/responses/Unauthentifiziert" }
        "403": { $ref: "#/components/responses/Verboten" }
        "412":
          description: Zwischenzeitlich geändert; neuerer Stand über GET erneut abrufen
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Fehler" }
        default: { $ref: "#/components/responses/Fehler" }
  /verwaltung/stammdaten:
    get:
      tags: [verwaltung]
      summary: Ferngepflegte Stammdaten zur Bearbeitung
      description: |
        Erfüllt ADMIN-F-180 und ADMIN-F-190. Liefert denselben Bestand wie
        `GET /stammdaten`, hier jedoch nur für die Rolle FSR-Redaktion und mit
        `ETag` für die optimistische Nebenläufigkeitskontrolle beim Ersetzen.
      responses:
        "200":
          description: Aktueller Stammdatenbestand
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Stammdaten" }
        "401": { $ref: "#/components/responses/Unauthentifiziert" }
        "403": { $ref: "#/components/responses/Verboten" }
        default: { $ref: "#/components/responses/Fehler" }
    put:
      tags: [verwaltung]
      summary: Stammdaten ersetzen
      description: |
        Erfüllt ADMIN-F-180, ADMIN-F-190 und ADMIN-F-200. Ersetzt den gesamten
        Bestand (Mensa-Liste, Raumliste, Links, Semestertermine, Ticket-Bildausschnitt).
        Der Kopf `If-Match` schützt vor dem Überschreiben einer zwischenzeitlichen
        Änderung durch eine andere Person.
      parameters:
        - name: If-Match
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Stammdaten" }
      responses:
        "200":
          description: Stammdaten gesetzt
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Stammdaten" }
        "401": { $ref: "#/components/responses/Unauthentifiziert" }
        "403": { $ref: "#/components/responses/Verboten" }
        "412":
          description: Zwischenzeitlich geändert; neuerer Stand über GET erneut abrufen (ADMIN-F-200)
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Fehler" }
        default: { $ref: "#/components/responses/Fehler" }
components:
  securitySchemes:
    authentik:
      type: openIdConnect
      openIdConnectUrl: https://auth.fsrfb4.de/application/o/fb4-app/.well-known/openid-configuration
      description: |
        OpenID Connect gegen die eigenbetriebene Authentik-Instanz (INT-012).
        Rollen kommen als Gruppen-Claim; das Backend wertet sie aus (API-F-250).
        Siehe decisions/0010-authentik-als-identitaetsanbieter.md.
...
  responses:
    Fehler:
      description: Fehler in einheitlichem Format (API-N-040)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Fehler" }
    Unauthentifiziert:
      description: Fehlende oder ungültige Authentifizierung
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Fehler" }
    Verboten:
      description: Rolle fehlt (ADMIN-F-010)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Fehler" }
...
  schemas:
    Fehler:
      type: object
      description: Einheitliches Fehlerformat nach RFC 9457, erfüllt API-N-040.
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string, description: Für Menschen lesbare Meldung }
        status: { type: integer }
        code: { type: string, description: Maschinenlesbarer Fehlercode }
        detail: { type: string }
...
    Stammdaten:
      type: object
      description: Erfüllt API-F-230.
      required: [mensen, raeume, links, semestertermine]
      properties:
        mensen:
          type: array
          items: { $ref: "#/components/schemas/Mensa" }
        raeume:
          type: array
          items: { $ref: "#/components/schemas/Raum" }
        links:
          type: array
          items: { $ref: "#/components/schemas/Link" }
        semestertermine: { $ref: "#/components/schemas/Semestertermine" }
        ticketBildausschnitt:
          $ref: "#/components/schemas/Bildausschnitt"
    Mensa:
      type: object
      required: [id, name, standardAuswahl, reihenfolge]
      properties:
        id: { type: string }
        name: { type: string }
        quelleId:
          type: string
          description: Kennung im Quellsystem INT-015.
        standardAuswahl: { type: boolean }
        reihenfolge: { type: integer }
        speiseplanUrl: { type: string, format: uri }
        oeffnungszeiten:
          type: array
          description: Ein Eintrag je Wochentag, Montag zuerst (MENSA-F-047).
          items:
            type: [string, "null"]
...
    Raumgroesse:
      type: string
      enum: [klein, mittel, gross]
    Raum:
      type: object
      required: [roomId, groesse, ekeyZugaenglich]
      properties:
        roomId: { type: string }
        groesse: { $ref: "#/components/schemas/Raumgroesse" }
        ekeyZugaenglich: { type: boolean }
        schliesszeit:
          type: [string, "null"]
          description: Gebäudeschließzeit als Endzeitpunkt, wenn kein Termin folgt (RAUM-F-130).
...
    Studiengang:
      type: object
      required: [kurzname, name, fachsemester]
      properties:
        kurzname: { type: string }
        name: { type: string }
        fachsemester:
          type: array
          items: { type: integer }
...
    Rolle:
      type: string
      enum: [fsr-redaktion, moderation]
    Rollenzuweisung:
      type: object
      required: [kontoId, rollen]
      properties:
        kontoId: { type: string }
        anzeigename: { type: string }
        rollen:
          type: array
          items: { $ref: "#/components/schemas/Rolle" }
    Laufweg:
      type: object
      required: [vonRoomId, nachRoomId, gewicht]
      properties:
        vonRoomId: { type: string }
        nachRoomId: { type: string }
        gewicht:
          type: number
          description: Fußweg in Minuten (RAUM, Abschnitt 13).
    Link:
      type: object
      required: [bezeichnung, url]
      properties:
        bezeichnung: { type: string }
        url: { type: string, format: uri }
        gruppe: { type: string }
        reihenfolge: { type: integer }
    Semestertermine:
      type: object
      properties:
        semesterBeginn: { type: string, format: date }
        semesterEnde: { type: string, format: date }
        naechsterWinterSemesterBeginn: { type: string, format: date }
        naechsterSommerSemesterBeginn: { type: string, format: date }
    Bildausschnitt:
      type: object
      description: Zuschnitt für die Ticketanzeige (TICKET-F-080).
      required: [links, oben, rechts, unten]
      properties:
        links: { type: integer }
        oben: { type: integer }
        rechts: { type: integer }
        unten: { type: integer }